# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dongjing server",
        "title": "gin-base-framework Server api",
        "termsOfService": "https://github.com/hexiaoyun128",
        "contact": {},
        "license": {},
        "version": "2.0"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登录"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"token\":\"Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjc1MzIxNjI5NTQsIk5hbWUiOiJhZG1pbiIsIlJvbGUiOm51bGwsIlVzZXJJZCI6MSwiSXNBZG1pbiI6dHJ1ZX0.YzPovX4xP6PPlZV9UGPKWgoLfGL8hnnC01j3L-k6f56mJuds7UDL--3Nts_P6RhuOQgzv7BL7hh6CJdSJdopjzE4A4HmEsq80_DN7cazuFE6gzA2ZfVLI7jslnWcJmVHVPTfu8_57NScdfxCDX_nFbbZUWjzDy7iT5L5zLXrBvg\",\"user\":{\"id\":1,\"name\":\"admin\",\"alias\":\"\",\"email\":\"admin@block.vc\",\"password\":\"\",\"roles\":[],\"openid\":\"\",\"active\":true,\"is_admin\":true}},\"message\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login/wechat": {
            "post": {
                "description": "用户微信登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登录"
                ],
                "summary": "用户微信登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "微信openid",
                        "name": "openid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户头像",
                        "name": "head",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"token\":\"Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjc1MzIxNjMwODMsIk5hbWUiOiJhZG1pbiIsIlJvbGUiOm51bGwsIlVzZXJJZCI6MiwiSXNBZG1pbiI6ZmFsc2V9.HZq5jBw4-ZQipQPnq0K7Ei0_LvaRXZGNgKqLoFnhV_vpfQupmddsDMZbiI_Yy0Zhd7J7AvRGDXMfVwW9-TidsDrux6-L4KQWIV0Mrlj4SXgW13HvMSXW0XzHYQBxiai61AeJx4VmQR84s2lI5hmKuiVOpsyOZAduJoO1K26b8X4\",\"user\":{\"id\":2,\"name\":\"admin\",\"alias\":\"\",\"email\":\"\",\"password\":\"\",\"roles\":[],\"openid\":\"admin\",\"active\":true,\"is_admin\":false}},\"message\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "用户退出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登录"
                ],
                "summary": "用户退出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"message\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}